MA-8489 - [imx8dv] pause/stat music, no sound from HDMI audio
On 8dv, if period_size is too small(176), when underrun,
the out_write_primary comsume 176 smaples quickly as only several us,
the producer (MonoPipe::write) is schduled at lower frequency,
and can't jump the loop. So the sii902x is always xrun.
One resolution is to calulate pcm_write interval, if too short for 10 times,
usleep 20ms, can fix the issue.
Here we enlarge period_size to avoid the issue.

Change-Id: I43234099e0cec80abedd96706bb237e041991002
Signed-off-by: fang hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/712
Reviewed-by: Winter Wang <wente.wang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1021