MA-8848 The preview is abnormal when the board is video recording module
and the result of taking picture is wrong.

There are ipu and pxp device on some boards. when board use camera hal the
code will determine pxp device exist or not, if pxp device exist the code
will ignore VIDIOC_S_INPUT. But VIDIOC_S_INPUT is needed on some boards. So
the csc should be based on the priority of different IP like IPU, PXP and GPU.
Modify the code to determine ipu device exist or not firstly.

Change-Id: Ic558297f42fffaa37f5ebae391a665e482b3a596
Signed-off-by: Wenbo.Zhu <wenbo.zhu_1@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1191
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>